Code cleanup, minus the extra item distribution testing
