+++ List/LocationList.py
@@ -148,9 +148,9 @@ class Locations:
             KH2Treasure(541, "Skillful Ring (Titan Cup)", LocationTypes=[locationType.OC, locationType.OCCups], InvalidChecks=[itemType.TROPHY,itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]).setLocationWeight(10),
             KH2Treasure(516, "Fatal Crest (Goddess of Fate Cup)", LocationTypes=[locationType.OC, locationType.OCCups], InvalidChecks=[itemType.TROPHY,itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]).setLocationWeight(10),
             KH2Treasure(517, "Orichalcum+ (Goddess of Fate Cup)", LocationTypes=[locationType.OC, locationType.OCCups], InvalidChecks=[itemType.TROPHY,itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]).setLocationWeight(10),
-            KH2Treasure(518, "Hades Cup Trophy (Paradox Cups)", LocationTypes=[locationType.OC, locationType.OCCups, locationType.OCParadoxCup], InvalidChecks=([] if trueRandom else [itemType.TORN_PAGE, itemType.FIRE, itemType.BLIZZARD, itemType.THUNDER, itemType.CURE, itemType.MAGNET, itemType.REFLECT])+[itemType.TROPHY, itemType.FORM, itemType.SUMMON, itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]),
-            KH2Treasure(546, "Zexion (AS) Book of Shadows", LocationTypes=[locationType.OC, locationType.AS], InvalidChecks=[itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]),
-            KH2Treasure(551, "Zexion (Data) Lost Illusion", LocationTypes=[locationType.OC, locationType.AS], InvalidChecks=[itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]),
+            KH2Treasure(518, "Hades Cup Trophy (Paradox Cups)", LocationTypes=[locationType.OC, locationType.OCCups, locationType.OCParadoxCup], InvalidChecks=([] if trueRandom else [itemType.TORN_PAGE, itemType.FIRE, itemType.BLIZZARD, itemType.THUNDER, itemType.CURE, itemType.MAGNET, itemType.REFLECT])+[itemType.TROPHY, itemType.FORM, itemType.SUMMON, itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]).setLocationWeight(10),
+            KH2Treasure(546, "Zexion (AS) Book of Shadows", LocationTypes=[locationType.OC, locationType.AS], InvalidChecks=[itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]).setLocationWeight(10),
+            KH2Treasure(551, "Zexion (Data) Lost Illusion", LocationTypes=[locationType.OC, locationType.AS], InvalidChecks=[itemType.GROWTH_ABILITY, itemType.SUPPORT_ABILITY, itemType.ACTION_ABILITY]).setLocationWeight(10),
 
             KH2Treasure(39, "Courtyard AP Boost", LocationTypes=[locationType.BC]).setLocationWeight(-10),
             KH2Treasure(40, "Courtyard Hi-Potion", LocationTypes=[locationType.BC]).setLocationWeight(-10),
