+++ Module/importantItems.py
@@ -1,0 +1,31 @@
+
+
+def getImportantChecks():
+    return [593,594,595, # proofs
+            21,22,23,24,87,88, #magic
+            26,27,29,31,563, # forms
+            32, #pages
+            159,160,25,383, #summons
+            415,416 #SC/OM
+    ]
+
+def getReports():
+    return [226,227,228,229,230,231,232,233,234,235,236,237,238]
+
+# subjective list of useful items used in item placement weighting
+def getUsefulItems():
+    return [535,362, #Munny Pouches
+            539, #Combo Master
+            393, #Finishing Plus
+            394, #Negative Combo
+            401, #Exp boost
+            541, #Light & Dark
+            271, #Hori Slash
+            264, #Slide Dash
+            559, #Flash Step
+            265, #Guard Break
+            266, #Explosion
+            560, #Aerial Dive
+            561, #Magnet Burst
+            198 #Trinity Limit
+            ]
